kernel: move event to events mod
