+++ src/events/mod.rs
@@ -22,6 +22,7 @@ pub(crate) mod bpf;
 pub(crate) mod file;
 
 pub(crate) mod ct;
+pub(crate) mod kernel;
 pub(crate) mod nft;
 pub(crate) mod ovs;
 pub(crate) mod skb;
