+++ src/process/cli/pcap.rs
@@ -23,8 +23,8 @@ use pcap_file::{
 
 use crate::{
     cli::*,
-    core::{kernel::Symbol, probe::kernel::KernelEvent},
-    events::{bpf::CommonEvent, file::FileEventsFactory, skb::SkbEvent, *},
+    core::kernel::Symbol,
+    events::{bpf::CommonEvent, file::FileEventsFactory, kernel::KernelEvent, skb::SkbEvent, *},
     helpers::signals::Running,
     module::{ModuleId, Modules},
 };
